// src/app/models/employee.model.ts
export interface EmployeeModel {
  readonly idx: string;
  readonly name: string;
  readonly salary: string;
}


export interface ApiModelEmp {
  readonly id: string;
  readonly employee_name: string;
  readonly employee_salary: string;
  readonly employee_age: string;
  readonly profile_image: string;
}